scheduler: isolate queue priority reprioritization
